// frontend-v2/src/app/core/api/models/produto-output.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProdutoPrecoOutput } from '../models/produto-preco-output';

/**
 * Dados de saida do produto
 */
export interface ProdutoOutput {

  /**
   * Descricao do produto
   */
  descricao?: string;

  /**
   * ID da categoria do produto
   */
  id_categoria: number;

  /**
   * ID do produto
   */
  id_produto: number;

  /**
   * Nome do produto
   */
  nome: string;

  /**
   * Opcoes de preco do produto
   */
  opcoes: Array<ProdutoPrecoOutput>;

  /**
   * Status do produto (ativo/inativo)
   */
  status: boolean;
}